+++ lib/ui/pages/friend/friend_list_page/widgets/friend_divider.dart
@@ -1,0 +1,17 @@
+import 'package:flutter/material.dart';
+
+class ListFriendDivider extends StatelessWidget {
+  const ListFriendDivider({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const Padding(
+      padding: EdgeInsets.symmetric(horizontal: 16),
+      child: Divider(
+        color: Colors.black,
+        thickness: 1,
+        height: 24,
+      ),
+    );
+  }
+}
