+++ lib/ui/pages/friend/friend_list_page/widgets/friend_tile.dart
@@ -1,0 +1,41 @@
+import 'package:flutter/material.dart';
+import 'package:tracky_flutter/ui/pages/friend/friend_detail_page/friend_detail_page.dart';
+
+class ListFriendTile extends StatelessWidget {
+  final String name;
+
+  const ListFriendTile({super.key, required this.name});
+
+  @override
+  Widget build(BuildContext context) {
+    return Padding(
+      padding: const EdgeInsets.symmetric(horizontal: 12),
+      child: ListTile(
+        leading: CircleAvatar(
+          backgroundColor: Colors.grey.shade400,
+          child: const Icon(Icons.person, color: Colors.white),
+        ),
+        title: Text(
+          name,
+          style: const TextStyle(
+            fontSize: 16,
+            color: Colors.black,
+          ),
+        ),
+        contentPadding: const EdgeInsets.symmetric(vertical: 4),
+        onTap: () {
+          Navigator.push(
+            context,
+            MaterialPageRoute(
+              builder: (_) => DetailFriendPage(
+                name: name,
+                email: '$name@nate.com',
+                isFriend: true,
+              ),
+            ),
+          );
+        },
+      ),
+    );
+  }
+}
